Extract local file check to its own extension
